whisper module added
