+++ app/chatcommands/normalmessage.js
